# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Currency Converter API
  description: API to convert currency between different currencies
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local Development Server
paths:
  /currency/supported-currencies:
    get:
      summary: Get list of supported currencies
      description: Fetches a list of available currency codes
      operationId: getSupportedCurrencies
      responses:
        "200":
          description: List of supported currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example: ["USD", "EUR", "INR", "GBP", "JPY", "CAD"]
        "500":
          description: Internal server error

  /currency/conversion:
    post:
      summary: Convert currency
      description: Converts an amount from one currency to another
      operationId: convertCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurrencyConversionInputDTO"
      responses:
        "200":
          description: Successfully converted the currency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrencyConversionResultDTO"
        "400":
          description: Invalid input parameters
        "500":
          description: Internal server error

components:
  schemas:
    CurrencyConversionInputDTO:
      type: object
      properties:
        sourceCurrency:
          type: string
          example: "USD"
        targetCurrency:
          type: string
          example: "INR"
        amount:
          type: number
          example: 100.0

    CurrencyConversionResultDTO:
      type: object
      properties:
        sourceCurrency:
          type: string
          example: "USD"
        destinationCurrency:
          type: string
          example: "INR"
        inputValue:
          type: number
          example: 100.0
        conversionResultValue:
          type: number
          example: 8350.0